Fix Ember.js computing the properties twice
Evidently, Ember.js will compute the property when you call this.get(PROPERTY),
even if you call it from INSIDE the computed property definition. This messes
some things up, and overall it's just bad for performance. But hey, at least it
isn't recursive enough to blow overflow the stack!

The fix is to simply give the computed property a new name instead of
overwriting the old one. In this case, since I'm making a "prettified" version
of the property for display in the views, I'll call it prettySalary. As a bonus,
this also leaves the salary property alone for math and stuff later.