Move job into separate template for reusability
By using a unique template that isn't directly tied to a controller, it can be
displayed in any other template using the {{partial}} helper. This means the
same template can be used in a job view, and in a comparison view, without
having to {{render}} the entire job view template in the comparison view.